requirements.txt
asgiref==3.8.1
colorama==0.4.6
dj-database-url==2.2.0
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
